agentChat_homework/backend/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-dotenv>=1.0.0
python-multipart>=0.0.6
autogen-agentchat==0.5.7
autogen-ext[openai]==0.5.7
autogen-core==0.5.7
pydantic>=2.10.0
sse-starlette>=1.8.2
